feat: add service settings view for Terraform
