+++ libs/domains/service-terraform/feature/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/hooks/use-create-terraform-service/use-create-terraform-service'
+export * from './lib/terraform-configuration-settings/terraform-configuration-settings'
 export * from './source-setting/source-setting'
