+++ libs/pages/application/src/lib/feature/page-settings-feature/page-settings-feature.tsx
@@ -15,6 +15,7 @@ import {
   APPLICATION_SETTINGS_PORT_URL,
   APPLICATION_SETTINGS_RESOURCES_URL,
   APPLICATION_SETTINGS_STORAGE_URL,
+  APPLICATION_SETTINGS_TERRAFORM_CONFIGURATION_URL,
   APPLICATION_SETTINGS_URL,
   APPLICATION_SETTINGS_VALUES_OVERRIDE_ARGUMENTS_URL,
   APPLICATION_SETTINGS_VALUES_OVERRIDE_FILE_URL,
@@ -129,6 +130,12 @@ export function PageSettingsFeature() {
     url: pathSettings + APPLICATION_SETTINGS_DANGER_ZONE_URL,
   }
 
+  const terraformConfigurationSetting = {
+    title: 'Terraform configuration',
+    iconName: 'sliders' as const,
+    url: pathSettings + APPLICATION_SETTINGS_TERRAFORM_CONFIGURATION_URL,
+  }
+
   const links = match(service)
     .returnType<NavigationLeftLinkProps[]>()
     .with({ serviceType: 'APPLICATION' }, () => [
@@ -163,6 +170,7 @@ export function PageSettingsFeature() {
     ])
     .with({ serviceType: 'TERRAFORM' }, () => [
       generalSettings,
+      terraformConfigurationSetting,
       resourcesSettings,
       deploymentRestrictionsSettings,
       advancedSettings,
