+++ libs/pages/application/src/lib/router/router.tsx
@@ -13,6 +13,7 @@ import {
   APPLICATION_SETTINGS_PORT_URL,
   APPLICATION_SETTINGS_RESOURCES_URL,
   APPLICATION_SETTINGS_STORAGE_URL,
+  APPLICATION_SETTINGS_TERRAFORM_CONFIGURATION_URL,
   APPLICATION_SETTINGS_URL,
   APPLICATION_SETTINGS_VALUES_OVERRIDE_ARGUMENTS_URL,
   APPLICATION_SETTINGS_VALUES_OVERRIDE_FILE_URL,
@@ -34,6 +35,7 @@ import { PageSettingsNetworkingFeature } from '../feature/page-settings-networki
 import PageSettingsPortsFeature from '../feature/page-settings-ports-feature/page-settings-ports-feature'
 import PageSettingsResourcesFeature from '../feature/page-settings-resources-feature/page-settings-resources-feature'
 import PageSettingsStorageFeature from '../feature/page-settings-storage-feature/page-settings-storage-feature'
+import PageSettingsTerraformConfigurationFeature from '../feature/page-settings-terraform-configuration-feature/page-settings-terraform-configuration-feature'
 import PageSettingsValuesOverrideArgumentsFeature from '../feature/page-settings-values-override-arguments-feature/page-settings-values-override-arguments-feature'
 import PageSettingsValuesOverrideFileFeature from '../feature/page-settings-values-override-file-feature/page-settings-values-override-file-feature'
 import PageVariablesFeature from '../feature/page-variables-feature/page-variables-feature'
@@ -113,5 +115,9 @@ export const ROUTER_APPLICATION_SETTINGS: Route[] = [
   {
     path: APPLICATION_SETTINGS_DANGER_ZONE_URL,
     component: <PageSettingsDangerZoneFeature />,
+  },
+  {
+    path: APPLICATION_SETTINGS_TERRAFORM_CONFIGURATION_URL,
+    component: <PageSettingsTerraformConfigurationFeature />,
   },
 ]
