+++ libs/shared/routes/src/lib/sub-router/application.router.ts
@@ -23,3 +23,4 @@ export const APPLICATION_SETTINGS_PORT_URL = '/port'
 export const APPLICATION_SETTINGS_DEPLOYMENT_RESTRICTIONS = '/deployment-restrictions'
 export const APPLICATION_SETTINGS_ADVANCED_SETTINGS_URL = '/advanced-settings'
 export const APPLICATION_SETTINGS_DANGER_ZONE_URL = '/danger-zone'
+export const APPLICATION_SETTINGS_TERRAFORM_CONFIGURATION_URL = '/terraform-configuration'
